On branch master  Your branch is up to date with 'origin/master'.
 Changes to be committed:
	modified:   auto-install.sh
	modified:   pre/partitioning.sh

+++ pre/partitioning.sh
@@ -12,8 +12,7 @@ check_returncode() {
         echo "OK"
         return 0
     else
-        echo -e "ERROR\n${ERRORTEXT}"
-        echo "=> Installation process aborted"
+        echo "${ERRORTEXT}" >&2
         exit $RETURN
     fi
 }
@@ -22,33 +21,37 @@ echo -e "\n$SUBHORIZONTALE"
 echo -e "\tPartitioning"
 echo -e "$SUBHORIZONTALE\n"
 
-echo -ne "Reading partition table...\t\t\t"
-PARTITIONTABLE=$(cat partition-table.conf 2>&1)
-check_returncode $? $PARTITIONTABLE
+if [ -f pre/partition-table.conf ]; then
 
-echo -ne "Applying partition instructions...\t\t\t"
-# sfdisk is a "scriptable" version of fdisk, it receives the "user input" from stdin.
-sfdisk /dev/sda 1>/dev/null 2>&1 <<- EOF
-    $PARTITIONTABLE
-EOF
-check_returncode $? "Partitioning with sfdisk failed"
+    echo -ne "Applying partition instructions...\t\t\t"
+    # sfdisk is a "scriptable" version of fdisk, it receives the "user input" from stdin.
+    sfdisk /dev/sda 1>/dev/null 2>&1 <<- EOF
+        $(cat pre/partition-table.conf)
+    EOF
+    check_returncode $? "Partitioning with sfdisk failed"
 
-echo -ne "Verifying Partition table...\t\t\t"
-# sfdisk -V: verifies the partition table of /dev/sda
-ERROR=$(sfdisk -V /dev/sda 2>&1 1>/dev/null)
-check_returncode $? $ERROR
+    echo -ne "Verifying Partition table...\t\t\t"
+    # sfdisk -V: verifies the partition table of /dev/sda
+    ERROR=$(sfdisk -V /dev/sda 2>&1 1>/dev/null)
+    check_returncode $? $ERROR
 
-## Subsection: Format the partitions
-echo -ne "Formatting root partition...\t\t\t"
-ERROR=$(mkfs.ext4 -q /dev/sda1 2>&1 1>/dev/null)
-check_returncode $? $ERROR
+    ## Subsection: Format the partitions
+    echo -ne "Formatting root partition...\t\t\t"
+    ERROR=$(mkfs.ext4 -q /dev/sda1 2>&1 1>/dev/null)
+    check_returncode $? $ERROR
 
-echo -ne "Initializing swap partition...\t\t\t"
-ERROR=$(mkswap /dev/sda2 2>&1 1>/dev/null)
-check_returncode $? $ERROR
+    echo -ne "Initializing swap partition...\t\t\t"
+    ERROR=$(mkswap /dev/sda2 2>&1 1>/dev/null)
+    check_returncode $? $ERROR
 
-echo -ne "Enabling swap partition...\t\t\t"
-ERROR=$(swapon /dev/sda2 2>&1 1>/dev/null)
-check_returncode $? $ERROR
+    echo -ne "Enabling swap partition...\t\t\t"
+    ERROR=$(swapon /dev/sda2 2>&1 1>/dev/null)
+    check_returncode $? $ERROR
+
+else
+    RETURN=$?
+    echo "pre/partition-table.conf not found" >&2
+    exit $RETURN
+fi
 
 exit 0
